% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd.R
\docType{data}
\name{cvd}
\alias{cvd}
\alias{protanomaly_cvd}
\alias{deutanomaly_cvd}
\alias{tritanomaly_cvd}
\title{Color vision deficiency (cvd) conversion tables}
\description{
Color vision deficiency (cvd) conversion tables
}
\keyword{datasets}
