% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_grid.R
\name{cvd_grid}
\alias{cvd_grid}
\title{Create a grid of different color-deficiency simulations of a plot}
\usage{
cvd_grid(plot = last_plot(), severity = 1)
}
\arguments{
\item{plot}{The plot to modify}

\item{severity}{The severity of the simulation, applied equally to all four cases}
}
\description{
Create a grid of different color-deficiency simulations of a plot
}
\examples{
cvd_grid(palette_plot(palette_OkabeIto, color_labels = FALSE))
}
