% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{desaturate}
\alias{desaturate}
\title{\code{desaturate} turns colors into their gray values}
\usage{
desaturate(col)
}
\arguments{
\item{col}{Vector of colors to convert}
}
\description{
\code{desaturate} turns colors into their gray values
}
