% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{deutan1}
\alias{deutan1}
\alias{deutan2}
\alias{deutan3}
\alias{deutan4}
\alias{deutan5}
\alias{deutan6}
\alias{deutan7}
\alias{deutan9}
\alias{deutan10}
\title{\code{deutan1} converts colors using the deutan model of colorblindness (low severity)}
\usage{
deutan1(colors, sev = "1")

deutan2(colors, sev = "2")

deutan3(colors, sev = "3")

deutan4(colors, sev = "4")

deutan5(colors, sev = "5")

deutan6(colors, sev = "6")

deutan7(colors, sev = "7")

deutan9(colors, sev = "9")

deutan10(colors, sev = "10")
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect}
}
\description{
\code{deutan1} converts colors using the deutan model of colorblindness (low severity)

\code{deutan2} converts colors using the deutan model of colorblindness

\code{deutan3} converts colors using the deutan model of colorblindness

\code{deutan4} converts colors using the deutan model of colorblindness

\code{deutan5} converts colors using the deutan model of colorblindness

\code{deutan6} converts colors using the deutan model of colorblindness

\code{deutan7} converts colors using the deutan model of colorblindness

\code{deutan8} converts colors using the deutan model of colorblindness

\code{deutan10} converts colors using the deutan model of colorblindness (high severity)
}
