% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{deutan_transform}
\alias{deutan_transform}
\title{Generate color transformation matrices using the deutan model of colorblindness}
\usage{
deutan_transform(sev = 1)
}
\arguments{
\item{sev}{Severity of the color vision defect, a number between 0 and 1}
}
\description{
Generate color transformation matrices using the deutan model of colorblindness
}
