% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{deutanomaly}
\alias{deutanomaly}
\title{\code{deutanomaly} converts colors using the deutanomaly model of colorblindness}
\usage{
deutanomaly(colors, sev = "3")
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect}
}
\description{
\code{deutanomaly} converts colors using the deutanomaly model of colorblindness
}
