% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_colors.R
\name{edit_colors}
\alias{edit_colors}
\title{Edit colors in existing plot or grid object}
\usage{
edit_colors(grob, colfun = passthrough, fillfun = colfun)
}
\arguments{
\item{colfun}{The function used to edit colors (\code{col} in \code{gpar} objects).}

\item{fillfun}{The function used to edit fill colors (\code{fill} in \code{gpar} objects).
By default the same as \code{colfun}.}
}
\description{
Edit colors in existing plot or grid object
}
