% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_swatches.R
\name{gg_color_gradient}
\alias{gg_color_gradient}
\title{Generate a continuous color gradient to be colored via ggplot2's \code{scale_fill_*} mechanism}
\usage{
gg_color_gradient(n = 200, ymargin = 0, title = NULL, title_size = 14,
  title_face = "plain", plot_margin = margin(title_size/2, title_size/2,
  title_size/2, title_size/2))
}
\arguments{
\item{n}{Number of distinct color slices to be drawn}

\item{ymargin}{Fraction of each swatch to be used as margin in the y direction}

\item{title}{Optional title to print above the color swatches. Can also be provided via \link{ggtitle}.}

\item{title_size}{Font size of the title}

\item{title_face}{Font face of the title}

\item{plot_margin}{Margin around the plot, specified via the ggplot2 function \link{margin}}
}
\description{
This function creates a ggplot2 object that can be directly styled by adding
a \code{scale_fill_*} expression. Thus, this function is particularly useful to visualize
existing ggplot2 color scales.
}
\examples{
gg_color_gradient() + ggtitle("Default ggplot2 continuous color scale")
}
