% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_swatches.R
\name{gg_color_swatches}
\alias{gg_color_swatches}
\title{Generate a set of color swatches to be colored via ggplot2's \code{scale_fill_*} mechanism}
\usage{
gg_color_swatches(n = 10, xmargin = 0.2, ymargin = 0, title = NULL,
  title_size = 14, title_family = "", title_face = "plain",
  plot_margin = margin(title_size/2, title_size/2, title_size/2,
  title_size/2))
}
\arguments{
\item{n}{Number of color swatches to generate}

\item{xmargin}{Fraction of each swatch to be used as margin in the x direction}

\item{ymargin}{Fraction of each swatch to be used as margin in the y direction}

\item{title}{Optional title to print above the color swatches. Can also be provided via \link{ggtitle}.}

\item{title_size}{Font size of the title}

\item{title_family}{Font family of the title}

\item{title_face}{Font face of the title}

\item{plot_margin}{Margin around the plot, specified via the ggplot2 function \link{margin}}
}
\description{
This function is similar to \link{palette_plot}, with two main differences: First, unlike
\link{palette_plot}, \code{gg_color_swatches} cannot label the colors with their hex code. Second, the colors don't
need to be provided as an argument, the resulting plot can be directly styled by adding
a \code{scale_fill_*} expression. Thus, this function is particularly useful to visualize
existing ggplot2 color scales.
}
\examples{
gg_color_swatches(8) + scale_fill_OkabeIto()
gg_color_swatches(10) + ggtitle("Default ggplot2 discrete color scale")
}
