% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_plot.R
\name{palette_plot}
\alias{palette_plot}
\title{Visualize a color palette as swatches of colors, possibly with labels printed on top}
\usage{
palette_plot(colors, label_size = 6, color_labels = TRUE)
}
\arguments{
\item{colors}{Vector of color names or hex codes}

\item{label_size}{Size of the color labels to be printed}

\item{color_labels}{Individual bool or vector of bools indicating for which colors the
color names should be printed on top of the color swatch.}
}
\description{
Visualize a color palette as swatches of colors, possibly with labels printed on top
}
\examples{
palette_plot(palette_OkabeIto)
palette_plot(c("red", "green", "yellow", "magenta"), color_labels = FALSE)
}
