% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{protan}
\alias{protan}
\title{Convert colors using the protan model of colorblindness}
\usage{
protan(colors, sev = 1)
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect, a number between 0 and 1}
}
\description{
Convert colors using the protan model of colorblindness
}
