% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{protan1}
\alias{protan1}
\alias{protan2}
\alias{protan3}
\alias{protan4}
\alias{protan5}
\alias{protan6}
\alias{protan7}
\alias{protan9}
\alias{protan10}
\title{\code{protan1} converts colors using the protan model of colorblindness (low severity)}
\usage{
protan1(colors, sev = "1")

protan2(colors, sev = "2")

protan3(colors, sev = "3")

protan4(colors, sev = "4")

protan5(colors, sev = "5")

protan6(colors, sev = "6")

protan7(colors, sev = "7")

protan9(colors, sev = "9")

protan10(colors, sev = "10")
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect}
}
\description{
\code{protan1} converts colors using the protan model of colorblindness (low severity)

\code{protan2} converts colors using the protan model of colorblindness

\code{protan3} converts colors using the protan model of colorblindness

\code{protan4} converts colors using the protan model of colorblindness

\code{protan5} converts colors using the protan model of colorblindness

\code{protan6} converts colors using the protan model of colorblindness

\code{protan7} converts colors using the protan model of colorblindness

\code{protan8} converts colors using the protan model of colorblindness

\code{protan10} converts colors using the protan model of colorblindness (high severity)
}
