% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scale_colorpalette_OkabeIto}
\alias{scale_colorpalette_OkabeIto}
\title{Color palette proposed by Okabe and Ito}
\format{An object of class \code{character} of length 8.}
\usage{
scale_colorpalette_OkabeIto
}
\description{
Two color palettes taken from the article "Color Universal Design" by Okabe and Ito, http://jfly.iam.u-tokyo.ac.jp/color/.
The variant \code{palette_OkabeIto} contains a gray color, while \code{palette_OkabeIto_black} contains black instead.
}
\keyword{datasets}
