% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_OkabeIto}
\alias{scale_colour_OkabeIto}
\alias{scale_color_OkabeIto}
\alias{scale_fill_OkabeIto}
\alias{scale_OkabeIto}
\title{Okabe-Ito color scale}
\arguments{
\item{...}{common discrete scale parameters: \code{name}, \code{breaks}, \code{labels}, \code{na.value}, \code{limits}, \code{guide}, and \code{aesthetics}.
See \link{discrete_scale} for more details.}

\item{use_black}{If \code{TRUE}, scale includes black, otherwise includes gray.}

\item{order}{Numeric vector listing the order in which the colors should be used. Default is 1:8.}

\item{darken}{Relative amount by which the scale should be darkened (for positive values) or lightened (for negatice
values).}

\item{alpha}{Alpha transparency level of the color. Default is no transparency.}
}
\description{
This is a color-blind friendly, qualitative scale with eight different colors. See \link{palette_OkabeIto} for details.
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() + scale_color_OkabeIto()
ggplot(iris, aes(Sepal.Length, fill = Species)) +
  geom_density(alpha = 0.7) + scale_fill_OkabeIto(order = c(1, 3, 5))

cowplot::plot_grid(
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = 0.6),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = 0.4),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = 0.2),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = 0),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = -0.2),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = -0.4),
  gg_color_swatches(8) + scale_fill_OkabeIto(darken = -0.6), ncol = 1)
}
