% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_colorblind.R
\name{simulate_colorblind}
\alias{simulate_colorblind}
\title{Apply colorblindness tranformation to a vector of hex values}
\usage{
simulate_colorblind(col, cvd)
}
\arguments{
\item{col}{The color vector}

\item{cvd}{A vector (3x3--9 values) specifying the color vision deficiency transform matrix}
}
\description{
Generate a vector of hex colors color blindness
}
\examples{
## doesn't currently work
# simulate_colorblind(c("#005000","#008600","#00BB00"),
#   tritanomaly['60'])
}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}
