% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_colorblind.R
\name{simulate_colorblind_hex}
\alias{simulate_colorblind_hex}
\title{Simulate colorblindness for a hex color given a cvd transform matrix}
\usage{
simulate_colorblind_hex(hex_str, cvd)
}
\arguments{
\item{hex_str}{The hex color string, e.g., '#FF0000'}

\item{cvd}{A vector (3x3--9 values) specifying the color vision deficiency transform matrix}
}
\description{
Generate a hex color simulating colorblindness.
Algorithm from http://www.daltonize.org/.
Code from https://github.com/muppetjones/remcolor/blob/master/R/SimulateColorBlind.R
}
\examples{
## doesn't currently work
# simulate_colorblind_hex("#01a203", tritanomaly_cvd['60'])

}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}
