% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cvd.R
\name{simulate_cvd}
\alias{simulate_cvd}
\title{Simulate color vision deficiency given a cvd transform matrix}
\usage{
simulate_cvd(col, cvd_transform)
}
\arguments{
\item{col}{A color or vector of colors e.g., "#FFA801" or "blue"}

\item{cvd_transform}{A 3x3 matrix specifying the color vision deficiency transform matrix}
}
\description{
This function takes valid R colors and transforms them according to a cvd transform matrix.
}
\examples{

 simulate_cvd(c("#005000","blue","#00BB00"),
 tritanomaly_cvd['6'][[1]])

}
\author{
Claire D. McWhite @clairemcwhite
}
\keyword{colorblind}
\keyword{colors,}
\keyword{cvd,}
