% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{tritan1}
\alias{tritan1}
\alias{tritan1}
\alias{tritan2}
\alias{tritan3}
\alias{tritan4}
\alias{tritan5}
\alias{tritan6}
\alias{tritan7}
\alias{tritan9}
\alias{tritan10}
\title{\code{tritan1} converts colors using the tritan model of colorblindness (low severity)}
\usage{
tritan1(colors, sev = "1")

tritan1(colors, sev = "1")

tritan2(colors, sev = "2")

tritan3(colors, sev = "3")

tritan4(colors, sev = "4")

tritan5(colors, sev = "5")

tritan6(colors, sev = "6")

tritan7(colors, sev = "7")

tritan9(colors, sev = "9")

tritan10(colors, sev = "10")
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect}
}
\description{
\code{tritan1} converts colors using the tritan model of colorblindness (low severity)

\code{tritan1} converts colors using the tritan model of colorblindness

\code{tritan2} converts colors using the tritan model of colorblindness

\code{tritan3} converts colors using the tritan model of colorblindness

\code{tritan4} converts colors using the tritan model of colorblindness

\code{tritan5} converts colors using the tritan model of colorblindness

\code{tritan6} converts colors using the tritan model of colorblindness

\code{tritan7} converts colors using the tritan model of colorblindness

\code{tritan8} converts colors using the tritan model of colorblindness

\code{tritan10} converts colors using the tritan model of colorblindness (high severity)
}
