% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cvd.R
\name{view_cvd}
\alias{view_cvd}
\title{Interactively view cvd simulations of a figure or plot}
\usage{
view_cvd(plot)
}
\arguments{
\item{plot}{The plot or grid object to view}
}
\description{
Interactively view cvd simulations of a figure or plot
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(iris, aes(Sepal.Length, fill=Species)) +
  geom_density(alpha = 0.7)
view_cvd(plot)
}
}
