% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{desaturate}
\alias{desaturate}
\title{Turn colors into their gray values}
\usage{
desaturate(col, amount = 1)
}
\arguments{
\item{col}{Vector of colors to convert}

\item{amount}{Degree of desaturation. 1=complete, 0=none.}
}
\description{
Turn colors into their gray values
}
