% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{desaturate}
\alias{desaturate}
\title{\code{desaturate} turns colors into their gray values}
\usage{
desaturate(col, rel_chroma = 0)
}
\arguments{
\item{col}{Vector of colors to convert}

\item{rel_chroma}{Degree of desaturation. 0=complete, 1=none.}
}
\description{
\code{desaturate} turns colors into their gray values
}
