% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{deutan}
\alias{deutan}
\alias{protan}
\alias{tritan}
\title{\code{deutan} converts colors using the deutan model of colorblindness}
\usage{
deutan(colors)

protan(colors)

tritan(colors)
}
\arguments{
\item{colors}{Vector of R colors to convert}
}
\description{
\code{deutan} converts colors using the deutan model of colorblindness

\code{protan} converts colors using the protan model of colorblindness

\code{tritan} converts colors using the tritan model of colorblindness
}
