% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{deutan}
\alias{deutan}
\title{Convert colors using the deutan model of colorblindness}
\usage{
deutan(colors, sev = 1)
}
\arguments{
\item{colors}{Vector of colors to convert}

\item{sev}{Severity of the color vision defect, a number between 0 and 1}
}
\description{
Convert colors using the deutan model of colorblindness
}
