% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_colorblind.R
\name{simulate_colorblind}
\alias{simulate_colorblind}
\title{Apply colorblindness tranformation to a vector of hex values}
\usage{
simulate_colorblind(col, cvd)
}
\arguments{
\item{col}{The color vector}

\item{cvd}{A 3x3 matrix specifying the color vision deficiency transform matrix}
}
\description{
Generate a vector of hex colors color blindness
}
\examples{

 simulate_colorblind(c("#005000","#008600","#00BB00"),
 tritanomaly_cvd['6'][[1]])
}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}
