<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_colorblind.R
\name{simulate_colorblind}
\alias{simulate_colorblind}
\title{Simulate colorblindness for a hex color given a cvd transform matrix}
\usage{
simulate_colorblind(col, cvd)
}
\arguments{
\item{col}{The hex color string, e.g., '#FF0000'}

\item{cvd}{A 3x3 matrix specifying the color vision deficiency transform matrix}
}
\description{
Generate a hex color simulating colorblindness.
Daltonization algorithm from http://www.daltonize.org/.
Aided by https://github.com/muppetjones/remcolor/blob/master/R/SimulateColorBlind.R
}
\examples{

 simulate_colorblind(c("#005000","#008600","#00BB00"),
 tritanomaly_cvd['6'][[1]])

}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_colorblind.R
\name{simulate_colorblind}
\alias{simulate_colorblind}
\title{Simulate colorblindness for a hex color given a cvd transform matrix}
\usage{
simulate_colorblind(col, cvd)
}
\arguments{
\item{col}{The hex color string, e.g., '#FF0000'}

\item{cvd}{A 3x3 matrix specifying the color vision deficiency transform matrix}
}
\description{
Generate a hex color simulating colorblindness.
Daltonization algorithm from http://www.daltonize.org/.
Aided by https://github.com/muppetjones/remcolor/blob/master/R/SimulateColorBlind.R
}
\examples{

 simulate_colorblind(c("#005000","#008600","#00BB00"),
 tritanomaly_cvd['6'])

}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}
>>>>>>> e5a1b7d61716a3ac1571d7ef8a76f0be8a6c88a2
